Add files via upload
System id or dummy data of system must be added to the database. To create an admin account the level of the registered user can be updated from student to admin. This project is mainly focused on creating a computer allocation web app for college computer facilities but can also work for cyber-cafes with minor tweeks.